Add new configuration parameter: dry_run
When "dry_run" is set to "true" do not modify node state or routing
tables. The script can be run safely and configuration parameters can be
tested.

Also, send all error messages to SMC and handle exceptions in a better
way.

Fix installer to wait until the running script is stopped before
installing and starting the new version during the policy push. The
installer sometimes did not start the new version because the old
version was still running.

Add more unit tests for configuration.